refactor: adjusting service name
